Refactoring: move ExecutingManagerUpdaterThread to a separate class
