+++ azkaban-common/src/main/java/azkaban/executor/ExecutorManagerUpdaterStage.java
@@ -1,0 +1,15 @@
+package azkaban.executor;
+
+public class ExecutorManagerUpdaterStage {
+
+  private String value = "not started";
+
+  public String get() {
+    return value;
+  }
+
+  public void set(String value) {
+    this.value = value;
+  }
+
+}
